Code Cleanup Isle 6
Changed IOC of StringBuilder, adding thread safety in case people want
to play with TAP in the future for Examples... or me, I like TAP, Okay
you got me, mostly for me. I am still deciding whether I should pass the
generic instance into the Behavior's Behave method. It is kind of nice
to have, because the behavior could be called from a separate containing
class to affect another object from it's parent. If the instance is not
needed pass null and don't use the param. O.Behave(null)